(MAINT) Update audio reference
Prior to this commit the audio for the introduction
pointed to the local resources folder; pointing to
the local resources folder for multimedia is not
yet supported, so the audio does not play.

This commit updates the reference to point to a
storage endpoint in Azure, which will enable the
audio to playback.